Add the ability to attach a script name to bench items.
The script names to into a second column. It is initially blank,
and can be set by the user interactively.

Also track the name in a bench_tree TCL array, where the script name
is used as an index into the array and the display name is the contents
of the array. This makes for a means of discovery for the named items.