Create a new tcl bench command as a hook for tcl
This lays out the code structure that I think I want for
getting access to the bench_tree items from TCL. The idea
is that there is a map from names to FitsbenchItem pointers
that gives custom commands a hook to get at images. From this
it is just a matter of implementing commands to manipulate
the FitsbenchItem objects.